requirements.txt
asgiref==3.3.1
Django==2.2.10
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
urllib3==1.26.3
